refactor: extract useFetchRooms hook and fix direct navigation to /rooms and /analytics #160
